Fixed remaining issues with SmartObject. Moved most common code to SmartObject class. Overridden most locations. Big warp totems.

+++ NewMachinesMod/Data/MachineInformation.cs
@@ -62,7 +62,7 @@ namespace Igorious.StardewValley.NewMachinesMod.Data
         {
             return new CraftingRecipeInformation
             {
-                ID = m.ID,
+                ID = (int)m.ID,
                 Name = m.Name,
                 IsBig = true,
                 Materials = m.Materials.Select(_ => new IngredientInfo((int)_.Key, _.Value)).ToList(),
